feat: 카드 없어도 레이아웃 유지되도록 수정
