+++ Projects/Features/Scene/GroupDetailScene/Vote/VoteCore.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2024 com.mashup.gabbangzip. All rights reserved.
 //
 
+import Common
 import ComposableArchitecture
 import Foundation
