+++ Projects/Features/Scene/GroupDetailScene/Vote/VoteSwipeView/VoteSwipeView.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2024 com.mashup.gabbangzip. All rights reserved.
 //
 
+import Common
 import ComposableArchitecture
 import NukeUI
 import SwiftUI
@@ -55,54 +56,6 @@ struct CardView: View {
   }
 }
 
-public enum SwipeDirection {
-  case left
-  case right
-}
-
-struct DraggableViewModifier: ViewModifier {
-  @Binding var isActive: (Bool, SwipeDirection)
-  @State private var translation: CGSize = .zero
-  
-  private let threshold: CGFloat = 100.0
-  
-  func body(content: Content) -> some View {
-    content
-      .offset(x: translation.width)
-      .rotationEffect(.degrees(Double(translation.width / 20)))
-      .gesture(
-        DragGesture()
-        .onChanged { value in
-          translation = CGSize(width: value.translation.width, height: 0)
-        }
-        .onEnded { value in
-          if abs(value.translation.width) > threshold {
-            isActive = (false, handleSwipeDirection(value.translation.width))
-          } else {
-            translation = .zero
-          }
-        }
-      )
-      .animation(.interactiveSpring(), value: translation)
-  }
-  
-  private func handleSwipeDirection(_ width: CGFloat) -> SwipeDirection {
-    width > 0 ? .right : .left
-  }
-}
-
-extension View {
-  func draggable(
-    isActive: Binding<(Bool, SwipeDirection)>
-  ) -> some View {
-    self.modifier(
-      DraggableViewModifier(
-        isActive: isActive
-      )
-    )
-  }
-}
-
 #Preview {
   VoteSwipeView(
     imageURLs: .constant(
@@ -111,7 +64,7 @@ extension View {
         URL(string: "https://i.namu.wiki/i/hq6niPhkN8EhXuIkCNx32AN614AxXcaxKQ1EnyFaHN41caJM7rPfkfppaGZNlpgmXWPbkD_MGTbmGE4_BOrIBg.webp")
       ]
     ),
-    swipeAction: { _,_  in }
+    swipeAction: { _, _  in }
   )
 }
